microservices/auth_service/requirements.txt
# microservices/auth_service/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
redis==4.6.0
PyJWT==2.8.0
Werkzeug==2.3.7
bcrypt==4.0.1
psutil==5.9.5